+++ cuda_core/cuda/core/_memory/_buffer.pyx
@@ -30,7 +30,7 @@ from cuda.core._utils.cuda_utils cimport HANDLE_RETURN, _parse_fill_value
 import sys
 from typing import TypeVar
 
-from cuda.core._pycompat import BufferProtocol
+from cuda.core._utils.pycompat import BufferProtocol
 from cuda.core._dlpack import classify_dl_device, make_py_capsule
 from cuda.core._device import Device
